refactor: handler 패키지와 parser 패키지를 webserver 밑으로 옮김
