+++ src/main/java/webserver/handler/ExceptionHandler.java
@@ -1,4 +1,4 @@
-package handler;
+package webserver.handler;
 
 import exception.GeneralException;
 import webserver.response.Response;
