+++ src/main/java/webserver/handler/GetRequestHandler.java
@@ -1,4 +1,4 @@
-package handler;
+package webserver.handler;
 
 import annotation.GetMapping;
 import annotation.RequestParam;
