+++ src/test/java/parser/RequestHeaderParserTest.java
@@ -3,6 +3,7 @@ package parser;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import webserver.header.RequestHeader;
+import webserver.parser.RequestHeaderParser;
 
 import java.io.*;
